first three examples
I wrote the first three examples for using Ansible. It uses the
SECRETHUB_USERNAME environment variable to be able to create a repo
inside a user's personal namespace.

The project is divided into maps, each

TODO:
- Show inject functionality, but I'd rather not have a
${ {{ username }}/repo/secret } template with secrethub interpolation
syntax wrapping a Jinja2 template for use in ansible.
- Use environment: SECRETHUB_CREDENTIAL={{service_credential}} in
favor of writing the credential to file.